requirements.txt
streamlit>=1.36.0
pymupdf>=1.24.4
llama-cpp-python>=0.2.84
